Global settings

Global settings

VegaLite.renderer — Function.

renderer()

show current rendering mode (svg or canvas)

renderer(::Symbol)

set rendering mode (svg or canvas)

source

VegaLite.actionlinks — Function.

actionlinks()::Bool

show if plots will have (true) or not (false) the action links displayed

actionlinks(::Bool)

indicate if actions links should be dislpayed under the plot

source

VegaLite.junoplotpane — Function.

junoplotpane()::Bool

when using Juno, show if plots will be rendered in plotpane or not

junoplotpane(::Bool)

set if plots should be rendered in Juno's plotpane or not

source