Global settings

Note

This section is outdated and does not reflect the latest API of the package.

Global settings

VegaLite.renderer — Function.

renderer()

show current rendering mode (svg or canvas)

renderer(::Symbol)

set rendering mode (svg or canvas)

source

VegaLite.actionlinks — Function.

actionlinks()::Bool

show if plots will have (true) or not (false) the action links displayed

actionlinks(::Bool)

indicate if actions links should be dislpayed under the plot

source

junoplotpane