Outputs

Output

On evaluation

Saving to a file

VegaLite.pdf — Function.

pdf(filename::AbstractString, v::VLSpec{:plot})

Save the plot $v$ as a pdf file with name $filename$.

source

VegaLite.png — Function.

png(filename::AbstractString, v::VLSpec{:plot})

Save the plot $v$ as a png file with name $filename$.

source

VegaLite.svg — Function.

svg(filename::AbstractString, v::VLSpec{:plot})

Save the plot $v$ as a svg file with name $filename$.

source

Or you can use a single saving functions that guesses the image format from the extension of the provided filename

VegaLite.savefig — Function.

savefig(filename::AbstractString, v::VLSpec{:plot})

Save the plot $v$ as a file with name $filename$. The file format will be picked based on the extension of the filename.

source